# docs/swagger/Appreservas-Swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Api Documentation",
    "description": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost:8080",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
    "/app/appointment/create": {
      "post": {
        "tags": [
          "appointment-controller"
        ],
        "summary": "createAppointment",
        "operationId": "createAppointmentUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/app/appointment/delete": {
      "delete": {
        "tags": [
          "appointment-controller"
        ],
        "summary": "deleteAppointment",
        "operationId": "deleteAppointmentUsingDELETE",
        "parameters": [
          {
            "name": "customerId",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "employeeId",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/app/appointment/update": {
      "put": {
        "tags": [
          "appointment-controller"
        ],
        "summary": "updateAppointment",
        "operationId": "updateAppointmentUsingPUT",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/app/appointments/getAll": {
      "get": {
        "tags": [
          "appointment-controller"
        ],
        "summary": "getAllAppointments",
        "operationId": "getAllAppointmentsUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Appointment"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/app/appointments/getAvailable": {
      "get": {
        "tags": [
          "appointment-controller"
        ],
        "summary": "getAvailableAppointments",
        "operationId": "getAvailableAppointmentsUsingGET",
        "parameters": [
          {
            "name": "taskId",
            "in": "query",
            "description": "taskId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "description": "startTime",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "endTime",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "date",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/app/getMyAppointment/{id}": {
      "get": {
        "tags": [
          "appointment-controller"
        ],
        "summary": "getMyAppointment",
        "operationId": "getMyAppointmentUsingGET",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Appointment"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Appointment": {
        "title": "Appointment",
        "type": "object",
        "properties": {
          "appointmentId": {
            "type": "integer",
            "format": "int64"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "date": {
            "type": "string"
          },
          "employee": {
            "$ref": "#/components/schemas/Employee"
          },
          "endTime": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "task": {
            "$ref": "#/components/schemas/Task"
          }
        }
      },
      "AppointmentDTO": {
        "title": "AppointmentDTO",
        "type": "object",
        "properties": {
          "customerId": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string"
          },
          "employeeId": {
            "type": "integer",
            "format": "int64"
          },
          "endTime": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "startTime": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "taskId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Branch": {
        "title": "Branch",
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "branchId": {
            "type": "integer",
            "format": "int64"
          },
          "closureTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "name": {
            "type": "string"
          },
          "openingTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "Customer": {
        "title": "Customer",
        "type": "object",
        "properties": {
          "customerId": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Employee": {
        "title": "Employee",
        "type": "object",
        "properties": {
          "branch": {
            "$ref": "#/components/schemas/Branch"
          },
          "employeeId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "task": {
            "$ref": "#/components/schemas/Task"
          }
        }
      },
...
      "LocalTime": {
        "title": "LocalTime",
        "type": "object",
        "properties": {
          "hour": {
            "type": "integer",
            "format": "int32"
          },
          "minute": {
            "type": "integer",
            "format": "int32"
          },
          "nano": {
            "type": "integer",
            "format": "int32"
          },
          "second": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Task": {
        "title": "Task",
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "employees": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Employee"
            }
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "taskId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}